BJS2-97290 refac CODE MISHA
