+++ src/main/java/school/faang/user_service/config/KafkaTopicConfig.java
@@ -8,10 +8,25 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ConfigurationProperties(prefix = "topics")
 public class KafkaTopicConfig {
-    private String notifications;
-    private String analytics;
-    private String achievements;
+    private NotificationsTopics notifications;
+    private AnalyticsTopics analytics;
+    private AchievementsTopics achievements;
     private String events;
+
+    @Data
+    public static class NotificationsTopics {
+        private String likeReceived;
+    }
+
+    @Data
+    public static class AnalyticsTopics {
+        private String profileView;
+    }
+
+    @Data
+    public static class AchievementsTopics {
+        private String goalCompleted;
+    }
 }
